Remove temporary folders created by the preprocessing or the train processing
