--- tests/fixtures/example.xyz
@@ -1,5 +1,0 @@
-5.829999923706054688e+00 -1.882999992370605469e+01 2.399999946355819702e-01 1.580000000000000000e+02 1.670000000000000000e+02 1.760000000000000000e+02 1.000000000000000000e+00
-5.829999923706054688e+00 -1.886000061035156250e+01 2.199999988079071045e-01 1.510000000000000000e+02 1.620000000000000000e+02 1.760000000000000000e+02 2.000000000000000000e+00
-5.820000171661376953e+00 -1.884000015258789062e+01 2.399999946355819702e-01 1.570000000000000000e+02 1.640000000000000000e+02 1.760000000000000000e+02 2.000000000000000000e+00
-5.829999923706054688e+00 -1.884000015258789062e+01 2.300000041723251343e-01 1.550000000000000000e+02 1.650000000000000000e+02 1.760000000000000000e+02 3.000000000000000000e+00
-5.820000171661376953e+00 -1.887999916076660156e+01 2.300000041723251343e-01 1.540000000000000000e+02 1.670000000000000000e+02 1.770000000000000000e+02 1.000000000000000000e+00
