+++ tests/test_preprocess.py
@@ -1,9 +1,13 @@
 import numpy as np
+from pytest import mark
 
 from kpconv_torch import preprocess
 from kpconv_torch.io import ply
+from kpconv_torch.datasets.common import PointCloudDataset
+from kpconv_torch.utils.config import load_config
 
 
+@mark.dependency()
 def test_preprocess(dataset_path):
     """The preprocessing produces .PLY files usable for further training process in 'original_ply'
     subfolder as well as subsampling material in 'input_{subsampling_coef}' subfolder (KDTree,
@@ -23,6 +27,7 @@ def test_preprocess(dataset_path):
             |_ Area_3.ply
             |_ Area_5.ply
     """
+
     preprocess.preprocess("tests/tests_config_S3DIS.yml", dataset_path)
     subsampling_coef = 0.03
     assert (dataset_path / f"input_{subsampling_coef:.3f}").exists()
@@ -52,3 +57,48 @@ def test_preprocess(dataset_path):
             }
         )
         assert len(np.unique(labels)) == expected_label_count
+
+
+@mark.dependency(depends=["test_preprocess"])
+def test_augmentation_transform(input_points_array, dataset_path):
+    """
+    Unit test for augmentation_transform
+    """
+    # To make the result predictable
+    np.random.seed(0)
+
+    # Load configuration
+    config = load_config("tests/tests_config_S3DIS.yml")
+
+    dataset = PointCloudDataset(
+        config,
+        dataset_path,
+        ignored_labels=np.array([]),
+        chosen_log=None,
+        infered_file=None,
+        task="train",
+    )
+    augmented_points, scale, var_r = dataset.augmentation_transform(input_points_array)
+
+    assert np.all(
+        (augmented_points * 100).astype(np.int32)
+        == [
+            [-83, 86, 20],
+            [-85, 82, 20],
+            [-88, 81, 20],
+            [11, 17, 116],
+            [12, 17, 115],
+            [12, 17, 114],
+        ]
+    )
+
+    assert np.all((scale * 100).astype(np.int32) == [-104, 102, 100])
+
+    assert np.all(
+        (var_r * 100).astype(np.int32)
+        == [
+            [-95, 30, 0],
+            [-30, -95, 0],
+            [0, 0, 100],
+        ]
+    )
